spiral.message
A message originated as a single line of one or more space-separated words, but it was encoded into an nxm  matrix as a clockwise spiral starting in the lower left-hand corner. The message is decoded spirally starting from the lower left-hand corner of the matrix and moving in the clockwise direction (i.e., up, right, down, left, up, right, etc.). From the starting position, the matrix is traversed in the clockwise direction, scanning characters and switching to the next clockwise direction each time a boundary is reached (i.e., an already-scanned character or the end of the matrix). Scanning characters is continued till all the matrix's characters are scanned into a single decoded string. The word separator for the decoded string is the hash mark (#). The total number of words in the decoded string is printed as the result.