Create CatchTheVenotoise.java
A wild venotoise has spawned in the middle of the street, and catchers are nearby! The task is to find who will catch the venotoise.The street is represented as a long, straight line. The i th catcher is located at position xi along a line, and the venotoise is located at position x.The venotoise is stationary, and the i th catcher is moving towards the venotoise at a speed of vi units per second. A catcher moving at a speed of v units per second can travel a distance of d units in exactly d/v seconds. The first catcher that makes it to the location of the venotoise catches it. If there isn't a unique "first catcher", that is, if there are two or more catchers that initially reach the venotoise at the exact same time, then the venotoise disappears, and no one gets the catch.